imagenet-trainer/requirements.txt
braceexpand
numpy
scipy
torch
torchvision
pytorch-lightning
tensorboard
matplotlib
jupyterlab
funny_words
bash_kernel
humanhash3
-e git://github.com/tmbdev/webdataset.git#egg=webdataset
